requirements.txt
anyio==3.7.1
APScheduler==3.10.4
argcomplete==3.1.2
arrow==1.2.3
bidict==0.22.1
cashews==6.3.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
filetype==1.2.0
flowery==0.0.1
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
KiramiBot==0.3.2
linkify-it-py==2.0.2
loguru==0.7.2
mango-odm==0.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdit-py-emoji==0.1.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
motor==3.3.1
msgpack==1.0.6
multidict==6.0.4
nonebot-adapter-onebot==2.3.0
nonebot2==2.1.0
packaging==23.1
Pillow==9.5.0
pipx==1.2.0
playwright==1.38.0
pydantic==1.10.12
pyee==9.0.4
Pygments==2.16.1
pygtrie==2.5.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.5.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tenacity==8.2.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
userpath==1.9.1
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
